Bricks colliding, needs improvement

+++ src/org/academiadecodigo/twoballs/gameobjects/Brick.java
@@ -20,6 +20,10 @@ public class Brick extends GameObject {
 
 
     public void deleteBrick() {
-        shape.delete();
+        durability--;
+        if (durability == 0) {
+            shape.delete();
+
+        }
     }
 }
